Start to move bot routes to separate files.
